Simplify BPartner load test and reorganize k6 framework
- Removed extracted/mixed data strategies from BPartner load test
- BPartner test now only uses generated synthetic data
- Moved BPartner-specific framework files (data-generators.js, http-client.js) to bpartner/ directory
- Cleaned up DataManager to focus purely on audit replay functionality
- Removed unused framework files from audit-replay/
- Simplified run.sh configuration (removed DATA_STRATEGY, EXISTING_RATIO, BP_BODY_FILES)
- Updated documentation to reflect simplified architecture

Benefits:
- Clear separation between BPartner test and audit-replay tool
- Each tool is now self-contained with its own dependencies
- Simpler configuration and easier to understand
