+++ e2e/perf/k6/audit-replay/replay-api-audit.js
@@ -23,7 +23,7 @@
 import http from 'k6/http';
 import { check, sleep } from 'k6';
 import { Counter, Gauge, Rate, Trend } from 'k6/metrics';
-import { DataManager } from './../framework/data-manager.js';
+import { DataManager } from './framework/data-manager.js';
 
 // ==================== Configuration ====================
