+++ e2e/perf/k6/bpartner/http-client.js
