--- e2e/perf/k6/framework/data-manager.js
@@ -1,168 +1,0 @@
-/*
- * #%L
- * master
- * %%
- * Copyright (C) 2025 metas GmbH
- * %%
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as
- * published by the Free Software Foundation, either version 2 of the
- * License, or (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public
- * License along with this program. If not, see
- * <http://www.gnu.org/licenses/gpl-2.0.html>.
- * #L%
- */
-
-import { SharedArray } from 'k6/data';
-import { randomItem } from 'https://jslib.k6.io/k6-utils/1.4.0/index.js';
-import { generateBPartnerUpsert } from './data-generators.js';
-
-export class DataManager {
-  constructor({ bpartnerFiles = [], auditDataFile = null } = {}) {
-    this.existingBPartnerUpserts = [];
-    for (const filePath of bpartnerFiles) {
-      const arr = new SharedArray(`bp-${filePath}`, () => {
-        try {
-          const data = JSON.parse(open(filePath));
-          return Array.isArray(data) ? data : [];
-        } catch (e) {
-          return [];
-        }
-      });
-      this.existingBPartnerUpserts.push(...arr);
-    }
-
-    // Load audit data if provided
-    this.auditRequests = [];
-    if (auditDataFile) {
-      this.auditRequests = new SharedArray('audit-requests', () => {
-        try {
-          const data = JSON.parse(open(auditDataFile));
-          return data.requests || [];
-        } catch (e) {
-          console.error(`Failed to load audit data from ${auditDataFile}: ${e.message}`);
-          return [];
-        }
-      });
-    }
-    this.auditIndex = 0;
-  }
-
-  getBPartnerUpsert(strategy = 'mixed', existingRatio = 0.7) {
-    const useExisting =
-      strategy === 'existing' || (strategy === 'mixed' && Math.random() < existingRatio);
-
-    const base =
-      useExisting && this.existingBPartnerUpserts.length > 0
-        ? deepClone(randomItem(this.existingBPartnerUpserts))
-        : generateBPartnerUpsert();
-
-    return this._makeUniqueIdentifiers(base);
-  }
-
-  extractBPartnerIdentifier(upsertPayload) {
-    try {
-      return upsertPayload.requestItems?.[0]?.bpartnerIdentifier || null;
-    } catch (e) {
-      return null;
-    }
-  }
-
-  _makeUniqueIdentifiers(payload) {
-    const uniqueSuffix = `vu${__VU}-it${__ITER}-${Date.now()}`;
-    const result = deepClone(payload);
-
-    if (Array.isArray(result.requestItems)) {
-      for (const item of result.requestItems) {
-        if (item.bpartnerIdentifier) {
-          item.bpartnerIdentifier = ensureSuffix(item.bpartnerIdentifier, uniqueSuffix);
-        }
-        if (Array.isArray(item.locations)) {
-          for (const loc of item.locations) {
-            if (loc.locationIdentifier) {
-              loc.locationIdentifier = ensureSuffix(loc.locationIdentifier, uniqueSuffix);
-            }
-          }
-        }
-        if (Array.isArray(item.contacts)) {
-          for (const ct of item.contacts) {
-            if (ct.contactIdentifier) {
-              ct.contactIdentifier = ensureSuffix(ct.contactIdentifier, uniqueSuffix);
-            }
-            if (ct.email) {
-              const [name, domain] = String(ct.email).split('@');
-              ct.email = `${name}+${uniqueSuffix}@${domain || 'example.com'}`;
-            }
-          }
-        }
-      }
-    }
-    return result;
-  }
-
-  /**
-   * Get the next audit request for replay
-   * @param {string} mode - 'sequential' or 'random'
-   * @returns {Object|null} - Audit request object or null if no data
-   */
-  getNextAuditRequest(mode = 'sequential') {
-    if (!this.auditRequests || this.auditRequests.length === 0) {
-      return null;
-    }
-
-    if (mode === 'random') {
-      return deepClone(randomItem(this.auditRequests));
-    }
-
-    // Sequential mode
-    const request = this.auditRequests[this.auditIndex % this.auditRequests.length];
-    this.auditIndex++;
-    return deepClone(request);
-  }
-
-  /**
-   * Get total count of loaded audit requests
-   * @returns {number}
-   */
-  getAuditRequestCount() {
-    return this.auditRequests ? this.auditRequests.length : 0;
-  }
-
-  /**
-   * Filter audit requests by method
-   * @param {string} method - HTTP method (GET, POST, PUT, DELETE, PATCH)
-   * @returns {Array} - Filtered audit requests
-   */
-  getAuditRequestsByMethod(method) {
-    if (!this.auditRequests) return [];
-    return this.auditRequests.filter(req => req.method === method);
-  }
-
-  /**
-   * Filter audit requests by path pattern
-   * @param {RegExp|string} pattern - Path pattern to match
-   * @returns {Array} - Filtered audit requests
-   */
-  getAuditRequestsByPath(pattern) {
-    if (!this.auditRequests) return [];
-    const regex = pattern instanceof RegExp ? pattern : new RegExp(pattern);
-    return this.auditRequests.filter(req => regex.test(req.path));
-  }
-}
-
-function deepClone(obj) {
-  return JSON.parse(JSON.stringify(obj));
-}
-
-function ensureSuffix(value, suffix) {
-  if (!value) return value;
-  // Avoid doubling the same suffix if the source data already had one
-  return `${String(value).replace(/-vu\d+-it\d+-\d+$/, '')}-${suffix}`;
-}
