+++ e2e/perf/k6/run.sh
@@ -31,10 +31,7 @@ DEBUG_PARAMS=""
 # Configurable via env vars with safe defaults
 export BASE_URL="${BASE_URL:-http://localhost:8282/api/v2}"
 export AUTH_TOKEN="${AUTH_TOKEN}"
-export DATA_STRATEGY="${DATA_STRATEGY:-generated}"   # 'existing' | 'generated' | 'mixed'
-export EXISTING_RATIO="${EXISTING_RATIO:-0.7}"   # used when DATA_STRATEGY='mixed'
 export ORG_CODE="${ORG_CODE:-001}"                  # need to be provided, unless set in the request-body
-export BP_BODY_FILES="${BP_BODY_FILES:-../test-data/extracted-bpartner-requests.json}"
 
 # If VUS and DURATION are not specified, k6 wil use a ramping-vus plan. See bpartner-load-test.js
 export VUS="${VUS:-100}" # (VU=concurrent virtual user)
@@ -42,16 +39,12 @@ export DURATION="${DURATION:-500s}"
 
 echo "Running k6 with:"
 echo "  BASE_URL=${BASE_URL}"
-echo "  DATA_STRATEGY=${DATA_STRATEGY}"
-echo "  EXISTING_RATIO=${EXISTING_RATIO}"
 echo "  ORG_CODE=${ORG_CODE}"
-echo "  BP_BODY_FILES=${BP_BODY_FILES}"
+echo "  VUS=${VUS}"
+echo "  DURATION=${DURATION}"
 
 k6 run ${DEBUG_PARAMS} \
   -e BASE_URL="${BASE_URL}" \
   -e AUTH_TOKEN="${AUTH_TOKEN}" \
-  -e DATA_STRATEGY="${DATA_STRATEGY}" \
-  -e EXISTING_RATIO="${EXISTING_RATIO}" \
   -e ORG_CODE="${ORG_CODE}" \
-  -e BP_BODY_FILES="${BP_BODY_FILES}" \
   "${SCRIPT_DIR}/bpartner/bpartner-load-test.js"
